% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_bgc.R
\name{get_BGCs_from_BGC_json}
\alias{get_BGCs_from_BGC_json}
\title{Get information from antismash output json file}
\usage{
get_BGCs_from_BGC_json(BGC_json)
}
\arguments{
\item{BGC_json}{antismash object from \code{read_antismash_json}}
}
\value{
data.frame
}
\description{
Get information from antismash output json file
}
