% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{get_big_slice_db}
\alias{get_big_slice_db}
\title{Get big-slice database}
\usage{
get_big_slice_db(big_slice_res, table = "bgc")
}
\arguments{
\item{big_slice_res}{big_slice object}

\item{table}{default "bgc"}
}
\value{
data.frame
}
\description{
Get big-slice database
}
