% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{get_mibig_db}
\alias{get_mibig_db}
\title{Get MiBIG database files}
\usage{
get_mibig_db(download_dir, file = NULL)
}
\arguments{
\item{download_dir}{where to save the download .tar.gz file?}

\item{file}{mibig_json_.*tar.gz from https://dl.secondarymetabolites.org/mibig/}
}
\value{
No value
}
\description{
Get MiBIG database files
}
