% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{get_report_df}
\alias{get_report_df}
\title{Get report dataframe from big_slice_res}
\usage{
get_report_df(big_slice_res, report, distance = 0.4)
}
\arguments{
\item{big_slice_res}{big_slice object}

\item{report}{report name}

\item{distance}{the threshold to define a bgc in a gcf, default 0.4.}
}
\value{
data.frame
}
\description{
Get report dataframe from big_slice_res
}
