% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{how_to_do_bgc}
\alias{how_to_do_bgc}
\title{How to use some softwares to analysis bgcs}
\usage{
how_to_do_bgc(step = c("antismash", "big-scape", "big-slice"))
}
\arguments{
\item{step}{"antismash", "big-scape", "big-slice"}
}
\value{
No value
}
\description{
How to use some softwares to analysis bgcs
}
