% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{mibig_db2df}
\alias{mibig_db2df}
\title{Convert MiBIG database to data frame}
\usage{
mibig_db2df(mibig_db)
}
\arguments{
\item{mibig_db}{mibig_db object from \code{load_mibig_db()}}
}
\value{
mibig_db_df
}
\description{
Convert MiBIG database to data frame
}
