% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{open_big_slice_website}
\alias{open_big_slice_website}
\title{Open BiG-SLiCE output website}
\usage{
open_big_slice_website(big_slice_res, port = 1234)
}
\arguments{
\item{big_slice_res}{big_slice object}

\item{port}{port}
}
\value{
No value
}
\description{
Open BiG-SLiCE output website
}
