% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_bgc.R
\name{plot_BGC}
\alias{plot_BGC}
\title{Plot a BGC}
\usage{
plot_BGC(
  BGC_json,
  region_id = NULL,
  show_locus_tag = TRUE,
  show_gene_functions = FALSE
)
}
\arguments{
\item{BGC_json}{antismash object from \code{read_antismash_json}}

\item{region_id}{region_id, default is NULL}

\item{show_locus_tag}{logical, default is TRUE}

\item{show_gene_functions}{logical, default is FALSE}
}
\value{
ggplot
}
\description{
Plot a BGC
}
