% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{plot_BGC_class}
\alias{plot_BGC_class}
\title{Plot BGC class}
\usage{
plot_BGC_class(big_scape_res, mode = 1, rm_mibig = FALSE, ...)
}
\arguments{
\item{big_scape_res}{big_scape object from \code{read_big_scape_dir()}}

\item{mode}{1~2, 1 for doughnut plot, 2 for sankey plot}

\item{rm_mibig}{logical, remove MIBiG BGCs from the plot}

\item{...}{additional parameters for \code{gghuan} or \code{my_sankey}, such as \code{topN=10}}
}
\value{
ggplot
}
\description{
Plot BGC class
}
