% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.antismash}
\alias{print.antismash}
\title{Print antismash}
\usage{
\method{print}{antismash}(x, ...)
}
\arguments{
\item{x}{antismash}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print antismash
}
