% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.big_scape}
\alias{print.big_scape}
\title{Print big_scape}
\usage{
\method{print}{big_scape}(x, ...)
}
\arguments{
\item{x}{big_scape}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print big_scape
}
