% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mibig_db}
\alias{print.mibig_db}
\title{Print mibig_db}
\usage{
\method{print}{mibig_db}(x, ...)
}
\arguments{
\item{x}{mibig_db}

\item{...}{Additional arguments}
}
\value{
No value
}
\description{
Print mibig_db
}
