% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_bgc.R
\name{read_antismash_json}
\alias{read_antismash_json}
\title{Read json file from antismash directory}
\usage{
read_antismash_json(BGC_json_file)
}
\arguments{
\item{BGC_json_file}{antismash output json file}
}
\value{
antismash object
}
\description{
Read json file from antismash directory
}
