% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{read_big_scape_dir}
\alias{read_big_scape_dir}
\title{Read files from bigscape directory}
\usage{
read_big_scape_dir(
  big_scape_dir = "2024-05-14_16-00-37_hybrids_glocal",
  cutoff = NULL,
  reassign_GCF = TRUE
)
}
\arguments{
\item{big_scape_dir}{Directory of bigscape output such as 2024-05-14_16-00-37_hybrids_glocal}

\item{cutoff}{default NULL, set this when you have multiple cutoffs}

\item{reassign_GCF}{default TRUE. When some BGCs were assigned into different GCFs ,reassign them into the GCF with the largest number of BGCs.}
}
\value{
big_scape object
}
\description{
Read files from bigscape directory
}
