% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{read_big_slice_dir}
\alias{read_big_slice_dir}
\title{Read BiG-SLiCE output dir}
\usage{
read_big_slice_dir(big_slice_dir)
}
\arguments{
\item{big_slice_dir}{path to the BiG-SLiCE output directory}
}
\value{
big_slice object
}
\description{
Read BiG-SLiCE output dir
}
