% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCF_net.R
\name{trans_net}
\alias{trans_net}
\title{Transform the network to metanet object}
\usage{
trans_net(big_scape_res, class = "NRPS")
}
\arguments{
\item{big_scape_res}{big_scape object from \code{read_big_scape_dir()}}

\item{class}{default "NRPS", set this to the class you want to extract}
}
\value{
metanet object
}
\description{
Transform the network to metanet object
}
