% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGCkit-package.R
\docType{package}
\name{BGCkit-package}
\alias{BGCkit}
\alias{BGCkit-package}
\title{BGCkit: A kit for biosynthetic gene cluster (BGC) analysis}
\description{
Study the evolutionary relationship between microbial genomes, especially the evolutionary patterns of specific functional genes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Asa12138/BGCkit}
  \item Report bugs at \url{https://github.com/Asa12138/BGCkit/issues}
}

}
\author{
\strong{Maintainer}: Chen Peng \email{pengchen2001@zju.edu.cn} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

}
\keyword{internal}
